Revert "Fix/crop calendar view refactor"
