+++ src/components/CropSelector/CropCalendarView.js
@@ -20,16 +20,20 @@ import { AcUnit, AddCircle, LocalFlorist, WbSunny } from '@mui/icons-material';
 import React, { Fragment, useContext, useEffect, useState } from 'react';
 import {
   allMonths,
+  CropImage,
   CustomStyles,
   flipCoverCropName,
+  getRating,
+  LightButton,
   sudoButtonStyle,
   sudoButtonStyleWithPadding,
-} from '../../../shared/constants';
-import { Context } from '../../../store/Store';
-import '../../../styles/cropCalendarViewComponent.scss';
-import CropDetailsModalComponent from '../CropDetailsModal';
-import CropLegendModal from '../CropLegendModal';
-import RenderCrops from './RenderCrops';
+  trimString,
+} from '../../shared/constants';
+import { Context } from '../../store/Store';
+import '../../styles/cropCalendarViewComponent.scss';
+import CropDetailsModalComponent from './CropDetailsModal';
+import CropLegendModal from './CropLegendModal';
+import CropSelectorCalendarView from './CropSelectorCalendarView';
 
 const growthIcon = {
   color: 'white',
@@ -39,23 +43,85 @@ const CropCalendarViewComponent = (props) => {
   const { activeCropData } = props;
   const { state, dispatch } = useContext(Context);
   const [legendModal, setLegendModal] = useState(false);
-  const [nameSortFlag, setNameSortFlag] = useState(true);
-  const [selectedCropsSortFlag, setSelectedCropsSortFlag] = useState(true);
-  const [modalOpen, setModalOpen] = useState(false);
-  const [modalData, setModalData] = useState([{}]);
-
-  const dispatchValue = (value, type = 'UPDATE_ACTIVE_CROP_DATA') => {
-    dispatch({
-      type: type,
-      data: {
-        value: value,
-      },
-    });
-  };
+  const selectedBtns = state.selectedCrops.map((crop) => {
+    return crop.id;
+  });
 
   const handleLegendModal = () => {
     setLegendModal(!legendModal);
   };
+
+  const hasGoalRatingTwoOrLess = (crop = []) => {
+    const { selectedGoals } = state;
+
+    return crop.inactive || selectedGoals.every((rating) => crop.fields[rating] <= 2);
+  };
+
+  const addCropToBasket = (cropId, cropName, btnId, cropData) => {
+    let selectedCrops = {};
+    let cropArray = [];
+    selectedCrops['id'] = cropId;
+    selectedCrops['cropName'] = cropName;
+    selectedCrops['btnId'] = btnId;
+    selectedCrops['data'] = cropData;
+    cropArray = selectedCrops;
+
+    // // check if crop id exists inside state, if yes then remove it
+    if (state.selectedCrops.length > 0) {
+      // DONE: Remove crop from basket
+      let removeIndex = state.selectedCrops
+        .map(function (item) {
+          return item.btnId;
+        })
+        .indexOf(`${btnId}`);
+      if (removeIndex === -1) {
+        // element not in array
+        dispatch({
+          type: 'SELECTED_CROPS_MODIFIER',
+          data: {
+            selectedCrops: [...state.selectedCrops, selectedCrops],
+            snackOpen: true,
+            snackMessage: `${cropName} Added`,
+          },
+        });
+      } else {
+        let selectedCropsCopy = state.selectedCrops;
+        selectedCropsCopy.splice(removeIndex, 1);
+        dispatch({
+          type: 'SELECTED_CROPS_MODIFIER',
+          data: {
+            selectedCrops: selectedCropsCopy,
+            snackOpen: true,
+            snackMessage: `${cropName} Removed`,
+          },
+        });
+      }
+    } else {
+      // DONE: add the selected crop to state and change the state, show snackbar
+
+      dispatch({
+        type: 'SELECTED_CROPS_MODIFIER',
+        data: {
+          selectedCrops: [cropArray],
+          snackOpen: true,
+          snackMessage: `${cropName} Added`,
+        },
+      });
+    }
+  };
+
+  const getAverageGoalRating = (selectedGoals, crop) => {
+    // get goal rating for each crop and calculate+render rating
+    let goalRating = 0;
+    selectedGoals.forEach((goal) => {
+      if (crop.fields[goal]) {
+        goalRating += crop.fields[goal];
+      }
+    });
+
+    return getRating(goalRating / selectedGoals.length);
+  };
+
   const [activeGrowthPeriodState, setActiveGrowthPeriodState] = useState(state.activeGrowthPeriod);
 
   useEffect(() => {
@@ -72,6 +138,7 @@ const CropCalendarViewComponent = (props) => {
   };
 
   const sortReset = (from = 'cropName') => {
+    // reset to default
     const { selectedGoals } = state;
     let activeCropDataShadow = props.activeCropData;
     selectedGoals
@@ -90,7 +157,12 @@ const CropCalendarViewComponent = (props) => {
         });
       });
 
-    dispatchValue(activeCropDataShadow);
+    dispatch({
+      type: 'UPDATE_ACTIVE_CROP_DATA',
+      data: {
+        value: activeCropDataShadow,
+      },
+    });
   };
   const sortCropsByName = () => {
     let activeCropDataShadow = props.activeCropData;
@@ -110,7 +182,12 @@ const CropCalendarViewComponent = (props) => {
           return firstCropName.localeCompare(secondCropName);
         });
 
-        dispatchValue(activeCropDataShadow);
+        dispatch({
+          type: 'UPDATE_ACTIVE_CROP_DATA',
+          data: {
+            value: activeCropDataShadow,
+          },
+        });
       }
     } else {
       if (activeCropDataShadow.length > 0) {
@@ -132,7 +209,12 @@ const CropCalendarViewComponent = (props) => {
           return 0;
         });
 
-        dispatchValue(activeCropDataShadow);
+        dispatch({
+          type: 'UPDATE_ACTIVE_CROP_DATA',
+          data: {
+            value: activeCropDataShadow,
+          },
+        });
       }
     }
 
@@ -163,15 +245,134 @@ const CropCalendarViewComponent = (props) => {
             }
           });
 
-          dispatchValue(newActiveShadow);
+          dispatch({
+            type: 'UPDATE_ACTIVE_CROP_DATA',
+            data: {
+              value: newActiveShadow,
+            },
+          });
         }
       }
     } else {
+      // sort back to original values
       sortReset('selectedCrops');
     }
     setSelectedCropsSortFlag(!selectedCropsSortFlag);
   };
+  const [nameSortFlag, setNameSortFlag] = useState(true);
+  const [selectedCropsSortFlag, setSelectedCropsSortFlag] = useState(true);
+  const [modalOpen, setModalOpen] = useState(false);
+  const [modalData, setModalData] = useState([{}]);
 
+  const RenderCrops = ({ cropData, active }) => {
+    return cropData
+      .filter((crop) => (active ? !hasGoalRatingTwoOrLess(crop) : hasGoalRatingTwoOrLess(crop)))
+      .map((crop, index) => {
+        if (crop.fields['Zone Decision'] === 'Include')
+          return (
+            <TableRow
+              key={`cropRow${index}`}
+              style={hasGoalRatingTwoOrLess(crop) ? { opacity: '0.2' } : {}}
+            >
+              <TableCell
+                className="calendarTableCell"
+                style={{
+                  paddingBottom: '0px',
+                }}
+              >
+                <div className="tdContainer d-flex justify-content-between flex-nowrap">
+                  {crop.fields['Image Data'] ? (
+                    <Button
+                      size="small"
+                      onClick={() => {
+                        setModalData(crop);
+                        setModalOpen(!modalOpen);
+                      }}
+                    >
+                      <CropImage
+                        view={'calendar'}
+                        present={true}
+                        src={
+                          crop.fields['Image Data']['Key Thumbnail']
+                            ? `/images/Cover Crop Photos/${crop.fields['Image Data']['Directory']}/${crop.fields['Image Data']['Key Thumbnail']}`
+                            : 'https://placehold.it/100x100'
+                        }
+                        alt={crop.fields['Cover Crop Name']}
+                      />
+                    </Button>
+                  ) : (
+                    <Button
+                      size="small"
+                      onClick={() => {
+                        setModalData(crop);
+                        setModalOpen(!modalOpen);
+                      }}
+                    >
+                      <CropImage view={'calendar'} present={false} />
+                    </Button>
+                  )}
+
+                  <Button
+                    size="small"
+                    onClick={() => {
+                      setModalData(crop);
+                      setModalOpen(!modalOpen);
+                    }}
+                  >
+                    {crop.fields['Cover Crop Name'] !== 'Sorghum-sudangrass'
+                      ? flipCoverCropName(crop.fields['Cover Crop Name'])
+                      : trimString(flipCoverCropName(crop.fields['Cover Crop Name']), 15)}
+                  </Button>
+                </div>
+              </TableCell>
+              {state.selectedGoals.length === 0 ? (
+                ''
+              ) : (
+                <TableCell
+                  style={{
+                    paddingBottom: '0px',
+                    textAlign: 'center',
+                  }}
+                >
+                  {getAverageGoalRating(state.selectedGoals, crop)}
+                </TableCell>
+              )}
+              <TableCell colSpan="12">
+                <CropSelectorCalendarView from="calendar" data={crop} />
+              </TableCell>
+
+              <TableCell
+                style={{
+                  paddingBottom: '0px',
+                }}
+              >
+                {' '}
+                <LightButton
+                  id={`cartBtn${index}`}
+                  style={{
+                    borderRadius: '0px',
+                    width: '130px',
+                  }}
+                  className={
+                    selectedBtns.includes(crop.fields.id) ? 'activeCartBtn' : 'inactiveCartBtn'
+                  }
+                  onClick={() => {
+                    addCropToBasket(
+                      crop.fields['id'],
+                      crop.fields['Cover Crop Name'],
+                      `cartBtn${index}`,
+                      crop.fields,
+                    );
+                  }}
+                >
+                  {selectedBtns.includes(crop.fields.id) ? 'ADDED' : 'ADD TO LIST'}
+                </LightButton>
+              </TableCell>
+            </TableRow>
+          );
+        else return <Fragment />;
+      });
+  };
   return (
     <Fragment>
       {state.ajaxInProgress ? (
@@ -355,7 +556,9 @@ const CropCalendarViewComponent = (props) => {
                 <TableCell style={{ width: '17%', borderRight: '5px solid white' }}>
                   <Button onClick={sortCropsByName}>COVER CROPS</Button>
                 </TableCell>
-                {state.selectedGoals.length > 0 && (
+                {state.selectedGoals.length === 0 ? (
+                  ''
+                ) : (
                   <TableCell style={{ width: '13%', borderRight: '5px solid white' }}>
                     <Button onClick={sortReset}>AVERAGE GOAL RATING</Button>
                   </TableCell>
@@ -391,21 +594,18 @@ const CropCalendarViewComponent = (props) => {
             </TableHead>
 
             <TableBody className="calendarTableBodyWrapper">
-              {activeCropData.length > 0 && (
+              {activeCropData.length > 0 ? (
                 <>
-                  <RenderCrops
-                    active={true}
-                    cropData={activeCropData}
-                    setModalOpen={setModalOpen}
-                    modalOpen={modalOpen}
-                    setModalData={setModalData}
-                  />
+                  <RenderCrops active={true} cropData={activeCropData} />
                   <RenderCrops active={false} cropData={activeCropData} />
                 </>
+              ) : (
+                ''
               )}
             </TableBody>
           </Table>
         </TableContainer>
+        // </div>
       )}
       <CropLegendModal
         legendModal={legendModal}
