+++ src/components/CropSelector/CropSelector.js
@@ -11,7 +11,7 @@ import React, { useContext, useEffect, useState } from 'react';
 import { Context } from '../../store/Store';
 import '../../styles/cropSelector.scss';
 import MyCoverCropList from '../MyCoverCropList/MyCoverCropList';
-import CropCalendarViewComponent from './CropCalendarView/CropCalendarView';
+import CropCalendarViewComponent from './CropCalendarView';
 import CropSidebarComponent from './CropSidebar';
 import CropTableComponent from './CropTable';
 import ReactGA from 'react-ga';
